+++ yash-env/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Env::errexit_is_applicable`
 - `SystemEx::fd_is_pipe`
 - `SystemEx::set_blocking`
+- `impl From<trap::Condition> for stack::Frame`
 
 ### Changed